requirements.txt
astroid==2.4.2
cffi==1.15.1
click==8.0.1
colorama==0.4.4
cryptography==40.0.1
Flask==2.0.1
gunicorn==20.1.0
isort==5.6.4
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
mccabe==0.6.1
Pillow==8.2.0
pycparser==2.21
pylint==2.6.0
six==1.15.0
toml==0.10.2
Werkzeug==2.0.1
wrapt==1.12.1
